% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric_factor.R
\name{as_numeric_factor}
\alias{as_numeric_factor}
\title{Factor to numeric via character}
\usage{
as_numeric_factor(x)
}
\arguments{
\item{x}{factor to be converted to numeric}
}
\description{
Shorthand to avoid the all-to-common bug of going straight from
factor to numeric.
}
\examples{
year = factor(1998:2004)
# bad
as.numeric(year)
# good
as.numeric(as.character(year))
# not much less typing, but no nesting (and it will auto-complete)
as_numeric_factor(year)
}

