% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_setter.R
\name{break_setter}
\alias{break_setter}
\title{Returns an appropriate function to pass to the \code{breaks}
argument of any \code{continuous_scale}}
\usage{
break_setter(n = 5)
}
\arguments{
\item{n}{suggested number of breaks (defaults to 5)}
}
\description{
Returns an appropriate function to pass to the \code{breaks}
argument of any \code{continuous_scale}
}
\examples{
\dontrun{
ggplot(mtcars, aes(x = disp, y = mpg)) +
geom_point() +
    scale_y_log10()

 ggplot(mtcars, aes(x = disp, y = mpg)) +
     geom_point() +
         scale_y_log10(breaks = break_setter())

 ggplot(mtcars, aes(x = disp, y = mpg)) +
     geom_point() +
     scale_y_log10(breaks = break_setter(10))
}
}

