% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce}
\alias{coalesce}
\title{Coalesce - take the first non-missing value}
\usage{
coalesce(...)
}
\arguments{
\item{...}{vectors to coalesce}
}
\description{
Coalesce - take the first non-missing value
}
\author{
Martin Morgan on Stack Overflow
}
\references{
\url{http://stackoverflow.com/a/19257945/903061}
}

