% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma_vector.R
\name{comma_vector}
\alias{comma_vector}
\title{Takes a vector and returns a comma-space separated string for printing.}
\usage{
comma_vector(x, and = TRUE, oxford = TRUE)
}
\arguments{
\item{x}{vector to paste together}

\item{and}{boolean, whether an "and" should be inserted before the last
element}

\item{oxford}{boolean, whether to include an "Oxford comma" (before the
"and")}
}
\description{
Takes a vector and returns a comma-space separated string for printing.
}

