% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operators.R
\name{\%btwn\%}
\alias{\%btwn\%}
\title{btwn operator}
\usage{
a \%btwn\% b
}
\arguments{
\item{a}{Sequence to test against min/max of b.}

\item{b}{Sequence for which min/max will be determined and tested against.}
}
\description{
Tests (element-wise) if the first argument falls within
the range of the second.
}
\examples{
1 \%btwn\% c(0, 5)
0:4 \%btwn\% c(0.5, 3, 3.5)

}

