% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_prep.R
\name{path_prep}
\alias{path_prep}
\title{Converts Windows filepath to R format}
\usage{
path_prep(path = "clipboard")
}
\arguments{
\item{path}{Defaults to clipboard.}
}
\description{
Keep input in the clipboard (or paste it in).
Returns a valid R filepath string.
}
\author{
Tyler Rinker (found on Stack Overflow).
}

