% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_g.r
\name{theme_clean}
\alias{theme_clean}
\title{A blank ggplot theme, borrowed from somebody on github}
\usage{
theme_clean(base_size = 12, ...)
}
\arguments{
\item{base_size}{defaults to 12, should be fine mostly}

\item{...}{additional arguments based to theme_gray}
}
\description{
A blank ggplot theme, borrowed from somebody on github
}

