% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_g.r
\name{theme_g}
\alias{theme_g}
\title{Nice theme for ggplot}
\usage{
theme_g(font = c("Eras Medium ITC"), slant = FALSE, gridlines = TRUE,
  expand.margin = FALSE, theme_color = "#3B6E8F", ...)
}
\arguments{
\item{font}{to use, defaults to "Eras Medium ITC".}

\item{slant}{boolean for whether x-axis labels should be slanted. (FALSE)}

\item{gridlines}{boolean for whether there should be any gridlines (TRUE)}

\item{expand.margin}{boolean for whether the right margin needs padding.
Useful if \code{slant}
is \code{TRUE} and the labels on the right side are lengthy.}

\item{theme_color}{Color for plot margin lines and facet strip backgrounds.
Defaults to a nice dark blue.}

\item{...}{additional arguments passed to theme_bw}
}
\description{
Based off of \code{theme_bw}, this frames the plot in dark blue,
uses filled in rectangles with white text for facet labels,
eliminates minor grid lines, and has a different default font.
There are also some other options such as gridlines and angles
for x-axis labels.
}

